Initial commit 🎉
This is a crate that I will use to learn the insights of a lot of different sorting algorithms.

Besides that the goal is to create the `Sort` and `SortStable` traits, which can be used to sort all kinds of types while always picking the fastest sorting algorithms for that particular type. This works by using the new `specialized` feature which allows to override generic trait implementations. This new feature allows to combine the power of traits, which are really generic and easy to use, and the power of specialized functionality, which often delivers performance that is magnitudes faster then a generic implantation.

Maybe this crate will eventually even become the sort crate (https://crates.io/crates/sort) which is currently empty and owned by the crates.io team.