Added a Sort trait and a generic implementation for it that now finally is able to use the best sorting algorithm
The problem so far was that I had no clue how to use specialization to resolve intersections between, i.e., T: CountingSort and T: RadixSort.
I now found a way, using the `specialization` crate. It also makes use of the `specialization` feature and provides a macro that allows using `if let` statements to check if a T implements a trait. Since this macro generates code that only relies on the type system, there's no runtime overhead.